tir: Remove unused structures in `scopes` and rename to `stack`
