+++ compiler/hash-semantics/src/passes/discovery/visitor.rs
@@ -8,7 +8,7 @@ use hash_ast::{
 use hash_reporting::macros::panic_on_span;
 use hash_storage::store::statics::SequenceStoreValue;
 use hash_tir::{
-    scopes::Stack,
+    stack::Stack,
     tir::{
         DataDef, FnDef, FnTy, ModDef, ModKind, ModMember, Node, NodeOrigin, SymbolId, Term,
         TupleTy, Ty,
